warnings fixed

+++ EntityManager/TileMap.cpp
@@ -147,8 +147,8 @@ ENTITYMANAGER_API void TileMap::CreateTiles(float textureWidth, float textureHei
 		tile.second->Release();
 	createdTiles.clear();
 
-	this->tileWidth = tileWidth;
-	this->tileHeight = tileHeight;
+	this->tileWidth = (int)tileWidth;
+	this->tileHeight = (int)tileHeight;
 
 	int tileCountX = (int)std::ceil(textureWidth / tileWidth);
 	int tileCountY = (int)std::ceil(textureHeight / tileHeight);
